mock 1inch swap payloads + removed old routers and tests
