+++ .husky/pre-push
@@ -1,4 +1,4 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
-# git fetch origin && HUSKY=0 git rebase origin/main && yarn test
+git fetch origin && HUSKY=0 git rebase origin/main && yarn test
