+++ contracts/interfaces/IArrakisV1Router.sol
@@ -2,40 +2,68 @@
 
 pragma solidity 0.8.4;
 
-import {
-    IArrakisVaultV1,
-    IArrakisV1RouterBase
-} from "./IArrakisV1RouterBase.sol";
+import {IArrakisVaultV1} from "./IArrakisVaultV1.sol";
+import {IGauge} from "./IGauge.sol";
 
-interface IArrakisV1Router is IArrakisV1RouterBase {
-    function rebalanceAndAddLiquidity(
-        IArrakisVaultV1 pool,
-        uint256 amount0In,
-        uint256 amount1In,
-        bool zeroForOne,
-        uint256 swapAmount,
-        uint160 swapThreshold,
-        uint256 amount0Min,
-        uint256 amount1Min,
-        address receiver
-    )
-        external
-        returns (
-            uint256 amount0,
-            uint256 amount1,
-            uint256 mintAmount
-        );
+struct AddLiquidityData {
+    // maximum amount of token0 to forward on mint
+    uint256 amount0Max;
+    // maximum amount of token1 to forward on mint
+    uint256 amount1Max;
+    // the minimum amount of token0 actually deposited (slippage protection)
+    uint256 amount0Min;
+    // the minimum amount of token1 actually deposited (slippage protection)
+    uint256 amount1Min;
+    // account to receive minted tokens
+    address receiver;
+    // bool indicating to use native ETH
+    bool useETH;
+    // address of gauge to stake tokens in
+    address gaugeAddress;
+}
 
-    function rebalanceAndAddLiquidityETH(
+struct MintData {
+    // amount of token0 to deposit
+    uint256 amount0In;
+    // amount of token1 to deposit
+    uint256 amount1In;
+    // amount of LP tokens to mint
+    uint256 mintAmount;
+}
+
+struct RemoveLiquidityData {
+    // amount of LP tokens to burn
+    uint256 burnAmount;
+    // minimum amount of token0 to receive
+    uint256 amount0Min;
+    // minimum amount of token1 to receive
+    uint256 amount1Min;
+    // address to receive underlying tokens
+    address payable receiver;
+    // bool indicating if user wants to receive in native ETH
+    bool receiveETH;
+    // address of gauge to unstake from
+    address gaugeAddress;
+}
+
+struct SwapData {
+    // max amount being swapped
+    uint256 amountInSwap;
+    // min amount received on swap
+    uint256 amountOutSwap;
+    // bool indicating swap direction
+    bool zeroForOne;
+    // address for swap calls
+    address swapRouter;
+    // payload for swap call
+    bytes swapPayload;
+}
+
+interface IArrakisV1Router {
+    function addLiquidity(
         IArrakisVaultV1 pool,
-        uint256 amount0In,
-        uint256 amount1In,
-        bool zeroForOne,
-        uint256 swapAmount,
-        uint160 swapThreshold,
-        uint256 amount0Min,
-        uint256 amount1Min,
-        address receiver
+        AddLiquidityData memory _addData,
+        MintData memory _mintData
     )
         external
         payable
@@ -43,5 +71,32 @@ interface IArrakisV1Router is IArrakisV1RouterBase {
             uint256 amount0,
             uint256 amount1,
             uint256 mintAmount
+        );
+
+    function removeLiquidity(
+        IArrakisVaultV1 pool,
+        RemoveLiquidityData memory _removeData
+    )
+        external
+        returns (
+            uint256 amount0,
+            uint256 amount1,
+            uint128 liquidityBurned
+        );
+
+    function swapAndAddLiquidity(
+        IArrakisVaultV1 pool,
+        AddLiquidityData memory _addData,
+        SwapData memory _swapData,
+        address payable userToRefund
+    )
+        external
+        payable
+        returns (
+            uint256 amount0,
+            uint256 amount1,
+            uint256 mintAmount,
+            uint256 amount0Diff,
+            uint256 amount1Diff
         );
 }
