--- contracts/interfaces/IArrakisV1RouterBase.sol
@@ -1,66 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0
-
-pragma solidity 0.8.4;
-
-import {IArrakisVaultV1} from "./IArrakisVaultV1.sol";
-
-interface IArrakisV1RouterBase {
-    function addLiquidity(
-        IArrakisVaultV1 pool,
-        uint256 amount0Max,
-        uint256 amount1Max,
-        uint256 amount0Min,
-        uint256 amount1Min,
-        address receiver
-    )
-        external
-        returns (
-            uint256 amount0,
-            uint256 amount1,
-            uint256 mintAmount
-        );
-
-    function addLiquidityETH(
-        IArrakisVaultV1 pool,
-        uint256 amount0Max,
-        uint256 amount1Max,
-        uint256 amount0Min,
-        uint256 amount1Min,
-        address receiver
-    )
-        external
-        payable
-        returns (
-            uint256 amount0,
-            uint256 amount1,
-            uint256 mintAmount
-        );
-
-    function removeLiquidity(
-        IArrakisVaultV1 pool,
-        uint256 burnAmount,
-        uint256 amount0Min,
-        uint256 amount1Min,
-        address receiver
-    )
-        external
-        returns (
-            uint256 amount0,
-            uint256 amount1,
-            uint128 liquidityBurned
-        );
-
-    function removeLiquidityETH(
-        IArrakisVaultV1 pool,
-        uint256 burnAmount,
-        uint256 amount0Min,
-        uint256 amount1Min,
-        address payable receiver
-    )
-        external
-        returns (
-            uint256 amount0,
-            uint256 amount1,
-            uint128 liquidityBurned
-        );
-}
