+++ src/oneInchApiIntegration.ts
@@ -1,10 +1,12 @@
 import fetch from "node-fetch";
 
+type OneInchDataType = { to: string; data: string };
+
 const approveTokenData = async (
   networkId: string,
   tokenAddress: string,
   amount: string
-): Promise<{ to: string; data: string }> => {
+): Promise<OneInchDataType> => {
   try {
     const apiResponse = (await (
       await fetch(
@@ -36,11 +38,11 @@ const swapTokenData = async (
   amount: string,
   fromAddress: string,
   slippage: string
-): Promise<{ to: string; data: string }> => {
+): Promise<OneInchDataType> => {
   try {
     const apiResponse = (await (
       await fetch(
-        `https://api.1inch.io/v4.0/${networkId}/swap?fromTokenAddress=${fromTokenAddress}&toTokenAddress=${toTokenAddress}&amount=${amount}&fromAddress=${fromAddress}&slippage=${slippage}&disableEstimate=true&allowPartialFill=false&protocols=UNISWAP_V1,UNISWAP_V2,SUSHI,MOONISWAP,BALANCER,COMPOUND,CURVE,CURVE_V2_SPELL_2_ASSET,CURVE_V2_SGT_2_ASSET,CURVE_V2_THRESHOLDNETWORK_2_ASSET,CHAI,OASIS,KYBER,AAVE,IEARN,BANCOR,PMM1,CREAMSWAP,SWERVE,BLACKHOLESWAP,DODO,DODO_V2,VALUELIQUID,SHELL,DEFISWAP,SAKESWAP,LUASWAP,MINISWAP,MSTABLE,PMM2,SYNTHETIX,AAVE_V2,ST_ETH,ONE_INCH_LP,ONE_INCH_LP_1_1,LINKSWAP,S_FINANCE,PSM,POWERINDEX,PMM3,XSIGMA,CREAM_LENDING,SMOOTHY_FINANCE,SADDLE,PMM4,KYBER_DMM,BALANCER_V2,SETH_WRAPPER,CURVE_V2,CURVE_V2_EURS_2_ASSET,CURVE_V2_EURT_2_ASSET,CURVE_V2_XAUT_2_ASSET,CURVE_V2_ETH_CRV,CURVE_V2_ETH_CVX,CONVERGENCE_X,ONE_INCH_LIMIT_ORDER,ONE_INCH_LIMIT_ORDER_V2,DFX_FINANCE,FIXED_FEE_SWAP,DXSWAP,CLIPPER,SHIBASWAP,UNIFI,PSM_PAX,WSTETH,DEFI_PLAZA,FIXED_FEE_SWAP_V3,SYNTHETIX_WRAPPER,SYNAPSE,CURVE_V2_YFI_2_ASSET,CURVE_V2_ETH_PAL,POOLTOGETHER,ETH_BANCOR_V3,ELASTICSWAP,BALANCER_V2_WRAPPER`
+        `https://api.1inch.io/v4.0/${networkId}/swap?fromTokenAddress=${fromTokenAddress}&toTokenAddress=${toTokenAddress}&amount=${amount}&fromAddress=${fromAddress}&slippage=${slippage}&disableEstimate=true&allowPartialFill=false&protocols=UNISWAP_V2`
       )
     ).json()) as unknown as {
       tx: {
@@ -52,7 +54,8 @@ const swapTokenData = async (
         gas: string;
       };
     };
-
+    console.log("swap - apiResponse.tx.to: ", apiResponse.tx.to);
+    console.log("swap - apiResponse.tx.data: ", apiResponse.tx.data);
     return {
       to: apiResponse.tx.to,
       data: apiResponse.tx.data,
@@ -74,12 +77,15 @@ const quote1Inch = async (
   try {
     const apiResponse = (await (
       await fetch(
-        `https://api.1inch.io/v4.0/${networkId}/quote?fromTokenAddress=${fromTokenAddress}&toTokenAddress=${toTokenAddress}&amount=${amount}&protocols=UNISWAP_V1,UNISWAP_V2,SUSHI,MOONISWAP,BALANCER,COMPOUND,CURVE,CURVE_V2_SPELL_2_ASSET,CURVE_V2_SGT_2_ASSET,CURVE_V2_THRESHOLDNETWORK_2_ASSET,CHAI,OASIS,KYBER,AAVE,IEARN,BANCOR,PMM1,CREAMSWAP,SWERVE,BLACKHOLESWAP,DODO,DODO_V2,VALUELIQUID,SHELL,DEFISWAP,SAKESWAP,LUASWAP,MINISWAP,MSTABLE,PMM2,SYNTHETIX,AAVE_V2,ST_ETH,ONE_INCH_LP,ONE_INCH_LP_1_1,LINKSWAP,S_FINANCE,PSM,POWERINDEX,PMM3,XSIGMA,CREAM_LENDING,SMOOTHY_FINANCE,SADDLE,PMM4,KYBER_DMM,BALANCER_V2,SETH_WRAPPER,CURVE_V2,CURVE_V2_EURS_2_ASSET,CURVE_V2_EURT_2_ASSET,CURVE_V2_XAUT_2_ASSET,CURVE_V2_ETH_CRV,CURVE_V2_ETH_CVX,CONVERGENCE_X,ONE_INCH_LIMIT_ORDER,ONE_INCH_LIMIT_ORDER_V2,DFX_FINANCE,FIXED_FEE_SWAP,DXSWAP,CLIPPER,SHIBASWAP,UNIFI,PSM_PAX,WSTETH,DEFI_PLAZA,FIXED_FEE_SWAP_V3,SYNTHETIX_WRAPPER,SYNAPSE,CURVE_V2_YFI_2_ASSET,CURVE_V2_ETH_PAL,POOLTOGETHER,ETH_BANCOR_V3,ELASTICSWAP,BALANCER_V2_WRAPPER`
+        `https://api.1inch.io/v4.0/${networkId}/quote?fromTokenAddress=${fromTokenAddress}&toTokenAddress=${toTokenAddress}&amount=${amount}&protocols=UNISWAP_V2`
       )
     ).json()) as unknown as {
       toTokenAmount: string;
     };
-
+    console.log(
+      "quote - apiResponse.toTokenAmount: ",
+      apiResponse.toTokenAmount
+    );
     return apiResponse.toTokenAmount;
   } catch (error) {
     console.log(
@@ -89,4 +95,101 @@ const quote1Inch = async (
   }
 };
 
-export { swapTokenData, quote1Inch, approveTokenData };
+// Because we pin a blockNumber in the tests,
+// using live 1inch api often causes as the price on current vs pinned block can vary
+// we store valid payloads from 1inch api for the pinned block here to be used in the tests
+type MockPayloadObj = {
+  swapIn: string;
+  swapOut: string;
+  payload: string;
+};
+
+type MockPayloadScenario = {
+  [index: string]: MockPayloadObj;
+};
+
+type MockPayloads = {
+  [vaultAddress: string]: MockPayloadScenario;
+};
+
+const mockPayloads: MockPayloads = {
+  "0xAbDDAfB225e10B90D798bB8A886238Fb835e2053": {
+    // use A,B and swap A for B + use A,B and swap A for B and stake
+    scenario1: {
+      swapIn: "1848075258604769953925",
+      swapOut: "1841470497",
+      payload:
+        "0x2e95b6c80000000000000000000000006b175474e89094c44da98b954eedeac495271d0f0000000000000000000000000000000000000000000000642f32402dbd7ac885000000000000000000000000000000000000000000000000000000006845adec0000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000100000000000000003b6d0340ae461ca67b15dc8dc81ce7615e0320da1a9ab8d5cfee7c08",
+    },
+
+    // use A,B and swap B for A + use A,B and swap B for A and stake
+    scenario2: {
+      swapIn: "7152021949",
+      swapOut: "7133497642501280506078",
+      payload:
+        "0x2e95b6c8000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb4800000000000000000000000000000000000000000000000000000001aa4b31bd00000000000000000000000000000000000000000000016f5f4eeb6a96866c060000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000180000000000000003b6d0340ae461ca67b15dc8dc81ce7615e0320da1a9ab8d5cfee7c08",
+    },
+
+    // use only A, swap A for B + use only A, swap A for B and stake
+    scenario3: {
+      swapIn: "259151912834101875000",
+      swapOut: "258233245",
+      payload:
+        "0x2e95b6c80000000000000000000000006b175474e89094c44da98b954eedeac495271d0f00000000000000000000000000000000000000000000000e0c748bad4ba31138000000000000000000000000000000000000000000000000000000000e9f4f6e0000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000100000000000000003b6d0340ae461ca67b15dc8dc81ce7615e0320da1a9ab8d5cfee7c08",
+    },
+
+    // use only B, swap B for A + use only B, swap B for A and stake
+    scenario4: {
+      swapIn: "741054440",
+      swapOut: "739221768198475351635",
+      payload:
+        "0x2e95b6c8000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48000000000000000000000000000000000000000000000000000000002c2b97e800000000000000000000000000000000000000000000002611d31f87f8f5b9350000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000180000000000000003b6d0340ae461ca67b15dc8dc81ce7615e0320da1a9ab8d5cfee7c08",
+    },
+  },
+  "0xa6c49FD13E50a30C65E6C8480aADA132011D0613": {
+    // use A,B and swap A for B + (... and stake) + (... with native ETH) + (... with native ETH and stake)
+    scenario1: {
+      swapIn: "9983580511",
+      swapOut: "5235085527190800873",
+      payload:
+        "0x2e95b6c8000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48000000000000000000000000000000000000000000000000000000025311595f0000000000000000000000000000000000000000000000004504d26c3c257c9d0000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000100000000000000003b6d0340b4e16d0168e52d35cacd2c6185b44281ec28c9dccfee7c08",
+    },
+
+    // use only A, swap A for B + (... and stake) + (... with native ETH) + (... with native ETH and stake)
+    scenario2: {
+      swapIn: "9986210283",
+      swapOut: "5236464330592196481",
+      payload:
+        "0x2e95b6c8000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb4800000000000000000000000000000000000000000000000000000002533979eb000000000000000000000000000000000000000000000000450979bc88c103200000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000100000000000000003b6d0340b4e16d0168e52d35cacd2c6185b44281ec28c9dccfee7c08",
+    },
+  },
+  "0xF920baf4fF3e803244C680FFdd468677f50C07B7": {
+    // same swap on all scenarios
+    scenario1: {
+      swapIn: "10000000000",
+      swapOut: "5243695698734024152",
+      payload:
+        "0x2e95b6c8000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb4800000000000000000000000000000000000000000000000000000002540be4000000000000000000000000000000000000000000000000004521e1c86d9ef8c00000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000100000000000000003b6d0340b4e16d0168e52d35cacd2c6185b44281ec28c9dccfee7c08",
+    },
+  },
+  "0xd14c67985F88368F66E71AD73ce97890F6636cdd": {
+    // same swap for all scenarios
+    scenario1: {
+      swapIn: "10000000000000000000",
+      swapOut: "18949515731",
+      payload:
+        "0x2e95b6c8000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20000000000000000000000000000000000000000000000008ac7230489e8000000000000000000000000000000000000000000000000000000000004310154880000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000180000000000000003b6d0340b4e16d0168e52d35cacd2c6185b44281ec28c9dccfee7c08",
+    },
+  },
+};
+
+export {
+  swapTokenData,
+  quote1Inch,
+  approveTokenData,
+  mockPayloads,
+  OneInchDataType,
+  MockPayloads,
+  MockPayloadScenario,
+  MockPayloadObj,
+};
